feat: Add save_path parameter to comparar_imagenes function
The comparar_imagenes function in ocr.py has been updated to include a new optional parameter, save_path. This parameter allows the user to specify a directory where the cropped images will be saved. If the save_path is provided, the function will save the cropped images with the corresponding file names. This change provides more flexibility and control over the output of the function.